# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StudentVoiceAgent",
    "version": "0.1.0",
    "description": "Emo Agent / StudentVoiceAgent 后端 REST API。前端通过 Vue API 封装调用这些接口；AI 能力由后端内部服务编排并通过 /api/chat、/api/resume、/api/town 等接口对外暴露。"
  },
  "paths": {
    "/api/chat/send": {
      "post": {
        "tags": [
          "对话"
        ],
        "summary": "Send",
        "operationId": "send_api_chat_send_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_api_chat_send_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/history": {
      "get": {
        "tags": [
          "对话"
        ],
        "summary": "History",
        "operationId": "history_api_chat_history_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryItem"
                  },
                  "title": "Response History Api Chat History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/growth": {
      "get": {
        "tags": [
          "成长中心"
        ],
        "summary": "Get Growth State",
        "operationId": "get_growth_state_api_growth_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthStateItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "成长中心"
        ],
        "summary": "Update Growth Profile",
        "operationId": "update_growth_profile_api_growth_put",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrowthProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthStateItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "成长中心"
        ],
        "summary": "Clear Growth State",
        "operationId": "clear_growth_state_api_growth_delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/growth/memories": {
      "post": {
        "tags": [
          "成长中心"
        ],
        "summary": "Create Growth Memory",
        "operationId": "create_growth_memory_api_growth_memories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrowthMemoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthMemoryItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/growth/memories/{memory_id}": {
      "delete": {
        "tags": [
          "成长中心"
        ],
        "summary": "Delete Growth Memory",
        "operationId": "delete_growth_memory_api_growth_memories__memory_id__delete",
        "parameters": [
          {
            "name": "memory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Memory Id"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_send_api_chat_send_post": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "audio": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "type": "object",
        "title": "Body_send_api_chat_send_post"
      },
...
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "assistant"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "user_text": {
            "type": "string",
            "title": "User Text",
            "default": ""
          },
          "emotion": {
            "$ref": "#/components/schemas/EmotionResult"
          },
          "tts_audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tts Audio Url"
          },
          "user_created_at": {
            "type": "string",
            "title": "User Created At"
          },
          "assistant_created_at": {
            "type": "string",
            "title": "Assistant Created At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "text",
          "emotion",
          "user_created_at",
          "assistant_created_at"
        ],
        "title": "ChatResponse"
      },
      "EmotionResult": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "audio_weight": {
            "type": "number",
            "title": "Audio Weight"
          },
          "text_weight": {
            "type": "number",
            "title": "Text Weight"
          }
        },
        "type": "object",
        "required": [
          "label",
          "confidence",
          "audio_weight",
          "text_weight"
        ],
        "title": "EmotionResult"
      },
      "GrowthMemoryCreate": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Session Id"
          },
          "category": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Category",
            "default": "目标"
          },
          "content": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "content"
        ],
        "title": "GrowthMemoryCreate"
      },
      "GrowthMemoryItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "content",
          "created_at"
        ],
        "title": "GrowthMemoryItem"
      },
      "GrowthProfileItem": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "current_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current State"
          },
          "focus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus"
          },
          "personality": {
            "type": "string",
            "title": "Personality",
            "default": "warm"
          },
          "weekly_goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Goal"
          },
          "setup_completed": {
            "type": "boolean",
            "title": "Setup Completed",
            "default": false
          },
          "private_mode": {
            "type": "boolean",
            "title": "Private Mode",
            "default": true
          },
          "anonymous_default": {
            "type": "boolean",
            "title": "Anonymous Default",
            "default": false
          },
          "crisis_guard": {
            "type": "boolean",
            "title": "Crisis Guard",
            "default": true
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "created_at",
          "updated_at"
        ],
        "title": "GrowthProfileItem"
      },
      "GrowthProfileUpdate": {
        "properties": {
          "session_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Session Id"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "current_state": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Current State"
          },
          "focus": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus"
          },
          "personality": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Personality"
          },
          "weekly_goal": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Goal"
          },
          "setup_completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setup Completed"
          },
          "private_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Mode"
          },
          "anonymous_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anonymous Default"
          },
          "crisis_guard": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crisis Guard"
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "GrowthProfileUpdate"
      },
      "GrowthStateItem": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/GrowthProfileItem"
          },
          "memories": {
            "items": {
              "$ref": "#/components/schemas/GrowthMemoryItem"
            },
            "type": "array",
            "title": "Memories"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "memories"
        ],
        "title": "GrowthStateItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "emotion_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion Label"
          },
          "emotion_conf": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotion Conf"
          },
          "tts_audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tts Audio Url"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content_type",
          "content",
          "emotion_label",
          "created_at"
        ],
        "title": "HistoryItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "本地开发环境"
    },
    {
      "url": "http://localhost:8000",
      "description": "本地备用地址"
    }
  ]
}